requirements.txt
telethon
python-dotenv
schedule
